requirements.txt
tweepy==4.14.0
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
# NLP Libraries
transformers==4.41.2
torch==2.3.1
# Backtesting Libraries (for later)
backtrader==1.9.76.123
yfinance==0.2.40
numpy==1.26.4
matplotlib==3.9.0
alpha-vantage==2.3.1
openai==1.35.7 